Use “ignoring application responses” for service menu item selection.
This works around a bug in Mavericks whereby System Events claims to be unable to click a menu item, but actually succeeds in doing so regardless. We just suppress the error message through the ignoring block in AppleScript, at the expense of reporting a true error.

I should really report this bug to Apple, not that I expect anything will be done about it given my recent history.