AutoHotKey-based display of microphone state.
We still don’t have a way to safely start and stop the state monitor, but hopefully we will get there.

Initially, I tried to do this by incorporating AutoHotKey spawning into the dictation server. This did work (not listening for changes but attaching to the keyboard toggle). However, it somehow interfered with the RPyC network listener.