Clean up Citrix Viewer usage.
Better handle the case where the clipboard does not get modified by a copy command, as is common when starting dictation and nothing is selected.  Previously, we always treated this as an error.

Now use “set” rather than “edit” so that clipboard transfers are much faster.  With no clipboard contents, now use “edit”, but primarily for its side effect of turning on the microphone; also bring VMware Fusion to the front (but don’t activate it).

In any case, when using “edit,” we are not dependent on the response time as in the common case it is not necessary to do more than turn the microphone on in order to start dictating.