Start a dictation RDP session in Jump Desktop.
Involves:

- starting the VM if it isn’t already running
- finding an alias to a Jump Desktop RDP session (.jump) file
- opening the file
- getting the window name out of the file so we know what to match
- closing the annoying Jump Desktop window
- moving the RDP window (requires a “Dictation” snapshot in Moom)

The latter two would not be necessary if Jump Desktop were a better Mac citizen, but it’s fast, maintained and bridges RTF clipboards, so I should not complain!

FWIW, I would not have spent an entire afternoon fighting with AppleScriptObjC to do this if I hadn’t been doing it partially as a learning experience.  The amount of time I spent to persist a selected file as an alias is mind-boggling.