Restart directly into an RDP session.
If the dictation server isn't running when you attempt to restart (or stop) it, don't worry.

Notify that we are restarting/stopping the virtual machine.

While restarting, wait for RDP such that the "Start Dictation via RDP" script does not attempt to start the virtual machine while it is being restarted.