Use WaitNotifier, a decorator which throttles polling processes and their resultant notifications.
Particularly useful for preventing the “waiting for RDP” notifications from piling up as the timeout for socket.create_connection doesn’t actually apply if nothing is listening on that socket.

In addition, 0.2 seconds is too quick. We can get away with throttling to as long as two or three seconds, but we need to trade off responsiveness of the script overall with not overwhelming the OS X notification system to the point that it starts counting rather than displaying individual notifications.