Try to avoid server hangs.
It seems that the most common cause (if not the only cause) of the dictation server hangs is natlink.natConnect() hanging. Previously, in an effort to coalesce multiple connections to natlink across a single server session, we always connected to natlink upon connection to the server. This was hanging, and in most cases, we did not even need to use natlink in that server session.

So, instead, connect lazily to natlink.