Properly handle toggling dictation when the buffer is full screen.
Seems like the only reason I ever actually hid the buffer app is to
test whether it is visible. Now appswitch can do this otherwise by
printing a pid, there is no need for the window to flash.

In addition, when hiding a 10.7+ full-screen window, the space would
switch and not restore when showing it. This was even worse.