Do a better job of finding the “Send Dictation” menu item.
Turns out that AppleScript doesn’t actually distinguish between singular and plural forms in some cases, and this was returning a list of lists when we were not expecting it. Of course, Cocoa Scripting did a horrendous job returning a useful error message, so this took me half an hour to figure out.

In other news, it is 2016.